refactor(env-variants): share schema/resolver across UI + daemon
Addresses Codex review on PR #1042. Collapses three duplicated variant
resolvers and the UI-side YAML schema validator into a single browser-safe
`variant-resolver.ts` in `@agor/core/config`, re-exported via `config/browser`
→ `@agor-live/client`.

- New `packages/core/src/config/variant-resolver.ts`: pure, fs-free module
  owning `resolveVariant`, `validateExtends`, `validateAgorYmlSchema`,
  `validateRepoEnvironment`, `parseAgorYmlString`, and `wrapV1AsV2`.
- `agor-yml.ts` becomes a thin Node-only wrapper (fs read/write) and routes
  YAML I/O through `@agor/core/yaml` instead of importing `js-yaml` directly.
- UI `environmentConfig.ts` drops its inlined `resolveVariant` copy and
  imports the shared one from `@agor-live/client`.
- `EnvironmentTab.tsx` replaces its ad-hoc `validateRepoYaml` with the shared
  `validateRepoEnvironment`, so UI save enforces the same invariants as file
  import (required `start`/`stop`, single-level extends, default-in-variants).
- `RepoRepository.deriveV1FromV2` uses shared `resolveVariant` instead of
  hand-rolled spread merging, and reuses `wrapV1AsV2` directly.
- `render-snapshot.ts` switches to `resolveVariantOrThrow` now that the pure
  resolver returns `null` on miss.
- Tighten `POST /repos/:id/export-agor-yml` to ADMIN (was MEMBER) to match
  the UI Export button gating and the Import/Edit roles.
- CodeEditor: replace fuzzy `algorithm.name.includes('dark')` detection with
  a canonical `isDark` boolean from `ThemeContext`; type `LANGUAGE_EXTENSIONS`
  with `typeof json` so both factories are constrained to `() => LanguageSupport`.
