+++ packages/core/src/config/browser.ts
@@ -9,3 +9,4 @@ export * from './constants';
 export * from './repo-list';
 export * from './repo-reference';
 export * from './types';
+export * from './variant-resolver';
